commit - refactoring
